+++ AudioKitSynthOne/AppDelegate.swift
@@ -16,14 +16,14 @@ import AppCenterCrashes
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     let conductor = Conductor.sharedInstance
+
     var window: UIWindow?
+
     private var launchOptions: [UIApplication.LaunchOptionsKey: Any]?
 
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
 
-        //print("launchOptions:\(String(describing: launchOptions))") // AKLog is not enabled at this point
         self.launchOptions = launchOptions
 
         // Never Sleep mode is false
@@ -65,18 +65,20 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func applicationDidEnterBackground(_ application: UIApplication) {
+
         conductor.checkIAAConnectionsEnterBackground()
     }
 
     
     func applicationWillEnterForeground(_ application: UIApplication) {
+
         conductor.checkIAAConnectionsEnterForeground()
     }
 
     func applicationWillResignActive(_ application: UIApplication) {
+
         // Prevent app from sleeping if never sleep is toggled on
         // Toggle back to sleep mode after 1 minutes
-        
         DispatchQueue.main.asyncAfter(deadline: .now() + 60) {
             UIApplication.shared.isIdleTimerDisabled = false
         }
@@ -85,6 +87,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func applicationWillTerminate(_ application: UIApplication) {
+
         // Called when the application is about to terminate.
         // Save data if appropriate. See also applicationDidEnterBackground:.
         conductor.stopEngine()
@@ -116,17 +119,17 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func canOpenURL(_ url: URL) -> Bool {
+        
         return true
     }
 }
 
-
 extension AppDelegate {
 
     public func applicationLaunchedWithURL() -> URL? {
+
         let launchUrl = self.launchOptions?[.url] as? URL
         self.launchOptions = nil
         return launchUrl
     }
-    
 }
